Fix for ROL-1492 "Change users profile and admin permissions feature does not work"
The problem is caused by the Acegi user-cache. The user's role is changed in the Roller DB, but Acegi is still hanging on to the old user object. So access to the Server Admin page is denied. 

There might be a way to flush the Acegi user cache, but I don't think there is much need for this cache and I'd rather not introduce any more Acegi dependencies. So I removed the cache from secrity.xml and that fixes the problem.

git-svn-id: https://svn.apache.org/repos/asf/roller/trunk@559865 13f79535-47bb-0310-9956-ffa450edef68