Fix for https://issues.apache.org/jira/browse/ROL-1899 "Media Files system does not work well for Java applets"
Roller's media file system was designed for images, videos and not for arbitrary uploads. File URLs are opaque, do not include directory path information or file extensions. There is a directory hierarchy that you can use to organize your files, but moving files around will not break links. That is all "by design".

If you really need to have file paths and extensions, you can enable a setting that allows you to edit the "originalPath" of a file. It's called original path because it is intended to support old Roller 4 style file URLs for a blog's old migrated uploads. It was not intended for general purpose use and therefore it is not editable by default.

This fix adds a configuration option that allows you to enable editing of the "original path" setting of a media file. This allows you to set an alternative full path, e.g. 'photos/picture1.jpg' for media files. All file URLs will use the old file upload base path of /resources/


git-svn-id: https://svn.apache.org/repos/asf/roller/trunk@1053756 13f79535-47bb-0310-9956-ffa450edef68